ENH: use env instead of envs on lume.yml files matching yml standards as Github Actions. (Old versions using envs will be compatibles also)
