+++ lume/config/step_config.py
@@ -1,23 +1,10 @@
-import os
 from typing import Any, Dict, List, Optional
 
-import yaml
 from pydantic import BaseModel
-from yaml.parser import ParserError
 
 from lume.config.check_list_or_str_item import check_list_or_str_item
 from lume.config.check_os_list_or_str_item import check_os_list_or_str_item
-
-
-def read_env_from_file(filename):
-    if not filename or not os.path.isfile(filename):
-        return {}
-    try:
-        with open(filename) as file:
-            envs = yaml.load(file, Loader=yaml.FullLoader)
-            return envs
-    except ParserError:
-        return {}
+from lume.config.get_envs import get_envs
 
 
 class StepConfig(BaseModel):
@@ -45,10 +32,7 @@ class StepConfig(BaseModel):
             setup_detach["run"] = check_list_or_str_item(
                 setup_detach, "run", required=True, suffix=" (setup_detach)"
             )
-        envs = kdict.get("envs", {})
-        envs_from_file = read_env_from_file(kdict.get("envs_file"))
-        envs.update(envs_from_file)
-
+        envs = get_envs(kdict)
         return StepConfig(
             run=run,
             cwd=kdict.get("cwd"),
